+++ instrumentation/github.com/gocql/gocql/otelgocql/go.mod
@@ -11,6 +11,5 @@ require (
 	go.opentelemetry.io/contrib v0.22.0
 	go.opentelemetry.io/otel v1.0.0-RC2
 	go.opentelemetry.io/otel/metric v0.22.0
-	go.opentelemetry.io/otel/oteltest v1.0.0-RC2
 	go.opentelemetry.io/otel/trace v1.0.0-RC2
 )
