+++ instrumentation/github.com/gocql/gocql/otelgocql/gocql.go
@@ -20,6 +20,7 @@ import (
 	"github.com/gocql/gocql"
 
 	otelcontrib "go.opentelemetry.io/contrib"
+	"go.opentelemetry.io/contrib/instrumentation/github.com/gocql/gocql/otelgocql/internal"
 	"go.opentelemetry.io/otel/trace"
 )
 
@@ -30,7 +31,7 @@ func NewSessionWithTracing(ctx context.Context, cluster *gocql.ClusterConfig, op
 	cfg := newConfig(options...)
 	instruments := newInstruments(cfg.meterProvider)
 	tracer := cfg.tracerProvider.Tracer(
-		instrumentationName,
+		internal.InstrumentationName,
 		trace.WithInstrumentationVersion(otelcontrib.SemVersion()),
 	)
 	cluster.QueryObserver = &OTelQueryObserver{
