+++ instrumentation/github.com/gocql/gocql/otelgocql/instrument.go
@@ -18,6 +18,7 @@ import (
 	"log"
 
 	otelcontrib "go.opentelemetry.io/contrib"
+	"go.opentelemetry.io/contrib/instrumentation/github.com/gocql/gocql/otelgocql/internal"
 	"go.opentelemetry.io/otel/metric"
 	"go.opentelemetry.io/otel/metric/unit"
 )
@@ -44,7 +45,7 @@ type instruments struct {
 // from the given provider p.
 func newInstruments(p metric.MeterProvider) *instruments {
 	meter := p.Meter(
-		instrumentationName,
+		internal.InstrumentationName,
 		metric.WithInstrumentationVersion(otelcontrib.SemVersion()),
 	)
 	instruments := &instruments{}
